# Conflicts:
#	BASD.Helper/Entity/EntityBase.vb
#	BASD.Helper/bin/Debug/BASD.Helper.dll
#	BASD.Helper/bin/Debug/BASD.Helper.pdb
#	BASD.Helper/obj/Debug/BASD.Helper.dll
#	BASD.Helper/obj/Debug/BASD.Helper.pdb
#	BASD.Helper/obj/Debug/BASD.Helper.vbproj.GenerateResource.Cache
#	BASD.Helper/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
#	EmployeeDirectory.suo
#	EmployeeDirectory/Data/Entity/Employee.vb
#	EmployeeDirectory/Services/EmpEditService.vb
#	EmployeeDirectory/obj/x86/Debug/DesignTimeResolveAssemblyReferences.cache
#	EmployeeDirectory/obj/x86/Debug/DesignTimeResolveAssemblyReferencesInput.cache
#	EmployeeDirectory/obj/x86/Debug/EmployeeDirectory.vbproj.FileListAbsolute.txt
#	EmployeeDirectory/obj/x86/Debug/EmployeeDirectory.vbproj.GenerateResource.Cache
#	EmployeeDirectory/obj/x86/Debug/EmployeeDirectory.vbprojResolveAssemblyReference.cache

+++ BASD.Helper/Entity/EntityBase.vb
@@ -1,8 +1,9 @@
 ﻿Public Class EntityBase
     Implements IEntity
 
-    Private _id As Int64
-    Public Property Id() As Int64 Implements IEntity.Id
+    Private _id As Integer
+    Public Property Id() As Integer Implements IEntity.Id
+
         Get
             Return _id
         End Get
